backend/doc_processor/requirements.txt
fastapi 
uvicorn 
PyPDF2==3.0.1 
python-docx==0.8.11 
pydantic-settings
python-multipart